requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==42.0.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.7
django-allauth==0.61.1
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.4
fontawesome==5.10.1.post1
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
magic-filter==1.0.12
multidict==6.0.4
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
pipenv==2023.10.3
platformdirs==3.11.0
psycopg==3.1.13
psycopg2==2.9.9
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.5
wikipedia==1.4.0
yarl==1.9.2
